frontend/flyout/files: duplicate controls at the top
